renamed the python package to qflexcirq
